[feat] 변경한 token 로직 적용
